+++ src/components/explore/landcover/charts/landCover2.js
