+++ src/components/explore/landcover/styles/LandCoverSelect.module.css
@@ -1,5 +1,5 @@
 .landCoverSelect {
-  width: 300px;
-  margin-top: 10px;
-  margin-bottom: 20px;
+    width: 300px;
+    margin-top: 10px;
+    margin-bottom: 20px;
 }
